Play a sound at a position
If we enqueue a command to play a sound following an object3D at some
point in the future, then we also have to check if that object3D is
still in the scene at the time we want to play the sound.

To avoid that complexity, we just save the positions of the avatars and
play a sound at each position. This will result in the wrong sound
effects when avatars are moving around or when they leave in between the
keypress and the audio, but most of the time it's mostly correct, and
that's good enough for a first version.